style: move common structs into separate file
